// frontend/src/components/chatWindow/chatWindow.module.scss.d.ts
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

declare const styles: {
  'chatHeader': string;
  'chatbotTitle': string;
  'chatbotHeader': string;
  'queriusLogo': string;
  'dropdownContainer': string;
  'usernameMenuItem': string;
  'menuButtonIcon': string;
  'userMenu': string;
  'userLogoutMenu': string;
  'dropdownContent': string;
  'userIcon': string;
  'logoutMenuItem': string;
  'icon': string;
  'currentTenantId': string;
  'verticalSeparator': string;
  'disabled': string;
  'chat-window': string;
  'flex-container': string;
  'settings-icon': string;
  'help-icon': string;
  'tooltipApiKey': string;
};
export default styles;